requirements.txt
Django==4.2.9
django-cors-headers==4.3.1
PyGithub==2.1.1
python-dotenv==1.0.0
click==8.1.7
django-rest-framework==0.1.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1